+++ src/main/java/org/elasticsearch/search/aggregations/bucket/multi/terms/StringTermsAggregator.java
@@ -62,7 +62,7 @@ public class StringTermsAggregator extends Aggregator {
     }
 
     @Override
-    public void collect(int doc, int owningBucketOrdinal) throws IOException {
+    public void collect(int doc, long owningBucketOrdinal) throws IOException {
         collector.collect(doc);
     }
 
@@ -78,7 +78,7 @@ public class StringTermsAggregator extends Aggregator {
     }
 
     @Override
-    public StringTerms buildAggregation(int owningBucketOrdinal) {
+    public StringTerms buildAggregation(long owningBucketOrdinal) {
         final BytesRefHash bytes = collector.bucketOrds;
         final LongArray counts = collector.counts;
         final int size = Math.min(bytes.size(), requiredSize);
