+++ src/main/java/org/elasticsearch/search/aggregations/bucket/single/missing/MissingAggregator.java
@@ -79,7 +79,7 @@ public class MissingAggregator extends SingleBucketAggregator {
         }
 
         @Override
-        protected MissingAggregator create(ValuesSource valuesSource, int expectedBucketsCount, AggregationContext aggregationContext, Aggregator parent) {
+        protected MissingAggregator create(ValuesSource valuesSource, long expectedBucketsCount, AggregationContext aggregationContext, Aggregator parent) {
             return new MissingAggregator(name, factories, valuesSource, aggregationContext, parent);
         }
     }
